Un-anti-alias the control track for the volcano
Inkscape has a bug in their PNG exporter (a ... RATHER outstanding one)
which causes it to anti-alias even when you ask it not to. Fortunately,
Pinta can import SVG, respects the anti-aliasing flag, and can then export
to .png. (So can Gimp, and reportedly many other image editing programs.)
So, export the level itself in inkscape for pretty smoothness, but use
some other SVG reading program to do the control map for hard edges.